chatbot_theme_identifier/backend/requirement.txt
fastapi
uvicorn
groq
langchain
faiss-cpu
pdfplumber
Pillow
pytesseract
streamlit
python-dotenv
chromadb
pandas